added Leo notes for optimizing extraction of control file from Debian package
Summary:

I've achieved a 60-80X improvement in the performance of extracting
control files from Debian binary packages, which I believe has a
significant aspect on many operations in fab (its done frequently).

While this improves the performance of a very particular operation in
the pool (importing of binary packages with the correct epoch), the
effort wouldn't have been justified if it was just that. The debinfo
library is my contribution to fab optimization.

I've clocked the old method used in fab at around 70-90
invocations per second: "ar -p %s control.tar.gz | zcat | tar -O -xf -
./control 2>/dev/null" % package

Its hard to neutralize the effects of disk buffering for benchmarking,
so for the purposes of comparative benchmarking I'll just take it for
granted.

I'm now achieving 4000 (for very large packages) - 8000 (for small
packages) invocations per second with caching, and 300-1400 without
caching.

The optimizations that made this possible:
1) I eliminated expensive execution of subcommands entirely. I
implemented extraction of the control file in pure Python.

2) I cache the control file in $HOME/.debinfo

3) the key to the control file is the md5sum some of
the control.tar.gz member of the Debian package instead of the entire
Debian package (saves a lot of IO)

Caching the control file wasn't very effective until I found a way to
efficiently calculate a unique key. The previous method of of hashing
the entire Debian package was not only expensive in terms of CPU, but
it was expensive in that reading in hundreds of megabytes worth of
files (instead of just the parts I need) would flush whatever was
already inside the disk buffers, negatively effecting the performance
of the rest of the system (e.g., fab).